create project section "sysex ISP"
This is the code that will be uploaded into RAM at 0x1000000 and that will eventually re-program the chip while in-service.
The new fimrware data (*.bin file) to be programmed will be appended to this binary at link time or later.
The whole thing then gets sysex encoded and then can be uploaded to the bridge